requirements.txt
torch
transformers
pandas
numpy
matplotlib
openpyxl
scikit-learn
tqdm
hazm
python-bidi
arabic-reshaper
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0